+++ src/js/shade.js
@@ -57,7 +57,8 @@
                 .addClass(this.options.cssClass)
                 .css({
                     background: this.options.color ? this.options.color : '',
-                    opacity: 0
+                    opacity: 0,
+                    '-webkit-tap-highlight-color': 'rgba(0,0,0,0)'
                 })
                 .hide()
                 .on(events.click, function() {