+++ tests/unit/options.js
@@ -2,7 +2,7 @@ define([
     'text!fixtures/shade.html',
     '$',
     'shade'
-], function(fixture, $, modalCenter) {
+], function(fixture, $) {
     var Shade;
     var element;
     var shade;
@@ -65,6 +65,13 @@ define([
 
                 assert.equal(shade.options.zIndex, 1);
                 assert.isNumber(shade.options.zIndex);
+            });
+
+            it('correctly defines cssClass', function() {
+                var shade = new Shade(element);
+
+                assert.equal(shade.options.cssClass, Shade.DEFAULTS.cssClass);
+                assert.isString(shade.options.cssClass);
             });
 
             it('correctly defines click', function() {
@@ -135,7 +142,7 @@ define([
                 assert.isNumber(shade.options.zIndex);
             });
 
-            it('correctly defines a custom class', function() {
+            it('correctly defines custom CSS class', function() {
                 shade = new Shade(element, { cssClass: 'c-class' });
 
                 assert.equal(shade.options.cssClass, 'c-class');