+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-1.2.0
-    - Added `destroy` function
 1.1.0
-    - Added option, `cssClass`, to specify custom CSS class
+	- Added `destroy` function
+    - Adds option 'cssClass' to define custom class on shade element
+    - Inlined CSS to remove the need to include stylesheet with one property
+    - Fixed tests, and added new tests
+    - Renamed grunt build-dist task to build
 1.0.5
     - Allows `click` to take `false` as an argument. This allows users to have Shade be fully blocking.